modules/elysium_local_kafka_to_spark_streaming/requirements.txt
kafka-python==2.3.0
py4j==0.10.9.9
pyspark==4.0.1